Updated biometric function Added settings icon
